restructured project by moving all frontend files into /client to enable deployment on Render
